% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_iso500.R
\name{get_iso500}
\alias{get_iso500}
\title{Get Data from iso500.org.tr}
\usage{
get_iso500(year, all = FALSE)
}
\arguments{
\item{year}{Year}

\item{all}{Download all data, default is FALSE}
}
\value{
A data tibble
}
\description{
Get Data from iso500.org.tr
}
\examples{
\dontrun{
get_iso500(2019)
}

}
