% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso500-package.R
\docType{package}
\name{iso500-package}
\alias{iso500}
\alias{iso500-package}
\title{iso500: Download data from iso500.org.tr}
\description{
Download data from iso500.org.tr.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/emraher/iso500}
  \item Report bugs at \url{https://github.com/emraher/iso500/issues}
}

}
\author{
\strong{Maintainer}: Emrah Er \email{eer@eremrah.com}

}
\keyword{internal}
