% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso500-package.R
\docType{package}
\name{iso500-package}
\alias{iso500}
\alias{iso500-package}
\title{iso500: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/emraher/iso500}
  \item Report bugs at \url{https://github.com/emraher/iso500/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
